add VotesEncodingVariants for parsing different formats from JSON;
